Moved the WriteError() methods to SystemUtil
Moved the WriteError() methods to SystemUtil and replaced Console.WriteLine()
with SystemUtil.WriteError().
