+++ src/app/SharpRaven/Data/SentryUserFactory.cs
@@ -31,6 +31,8 @@
 using System;
 using System.Security.Principal;
 
+using SharpRaven.Utilities;
+
 namespace SharpRaven.Data
 {
     /// <summary>
@@ -84,7 +86,7 @@ namespace SharpRaven.Data
             }
             catch (Exception exception)
             {
-                Console.WriteLine(exception);
+                SystemUtil.WriteError(exception);
             }
 
             return null;
@@ -99,7 +101,7 @@ namespace SharpRaven.Data
             }
             catch (Exception exception)
             {
-                Console.WriteLine(exception);
+                SystemUtil.WriteError(exception);
             }
 
             return null;
